requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
aioredis==1.3.1
aiosqlite==0.16.1
appdirs==1.4.3
asgiref==3.3.4
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
Babel==2.9.1
CacheControl==0.12.6
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-jazzmin==2.4.7
django-modeltranslation==0.17.1
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
iso8601==0.1.14
lockfile==0.12.2
msgpack==0.6.2
multidict==5.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
psycopg2==2.8.6
pyparsing==2.4.6
pypika-tortoise==0.1.1
pyTelegramBotAPI==3.7.9
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
tortoise-orm==0.17.3
typing-extensions==3.10.0.0
urllib3==1.25.8
webencodings==0.5.1
yarl==1.6.3
